requirements.txt
Django==2.0.3
django-bootstrap-datepicker-plus==2.0.0
Pillow==5.0.0
python-datauri==0.2.6
python-dateutil==2.7.0
pytz==2018.3
six==1.11.0
